pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "openai>=1.68.2",
    "python-telegram-bot>=22.0",
    "vllm>=0.8.1",
]
